LayoutTests/fast/dom/NodeIterator/NodeIterator-leak-document-expected.txt

This test asserts that document doesn't leak when a NodeFilter callback referencing the document is created.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS endingDocumentCount - startingDocumentCount < testingCount * 0.8 is true
PASS successfullyParsed is true

TEST COMPLETE
